fixed cyclic dependencies
