+++ src/super_gradients/training/models/segmentation_models/registry.py
@@ -1,4 +1,4 @@
-from super_gradients.training.models.segmentation_models.stdc import STDCBackbone
+from super_gradients.training.models.segmentation_models.stdc.stdc_encoder import STDCBackbone
 from super_gradients.training.models.segmentation_models.unet.unet_encoder import UNetBackboneBase
 
 SEGMENTATION_BACKBONES = dict(
